expunge legacy code from polkadot-network
